Update patch set 1
Patch Set 1:

> Looks good, my only comment would be that the role is too vague.
 > Maybe the role should be force_host? Also this would be a global
 > role so who would this role be assigned too?

Yeah I called it force_host to start off with but then I thought we might want to expand it to work for things like force_delete as well... happy to change though?

To start off with we want to give the role to the CVL and MCC projects, and probably to the various *porting projects (as local admins are usually members of these and this feature would be useful to them).

Patch-set: 1
Reviewer: Gerrit User 1000082 <1000082@f182a8af-62ca-4ceb-ab81-14f8fec9c6c5>
Label: Verified=0