Update patch set 1
Patch Set 1: -Code-Review

> What's a good default for max_files?

we set max_files=32768 but this could depend on no of VMs on the hypervisor, volumes attach to the VM as well as OSDs and PGs in cinder volume ceph cluster. Could this be set in hiera? 

We noticed the issue when particular VM has multiple volumes attached and when new OSDs are added into cinder volume ceph cluster.

http://ceph-users.ceph.narkive.com/baW6rD2y/nfs-interaction-with-rbd
http://lists.ceph.com/pipermail/ceph-users-ceph.com/2013-December/026187.html

Patch-set: 1