Update patch set 1
Patch Set 1:

> > What's a good default for max_files?
 > 
 > we set max_files=32768 but this could depend on no of VMs on the
 > hypervisor, volumes attach to the VM as well as OSDs and PGs in
 > cinder volume ceph cluster. Could this be set in hiera?
 > 
 > We noticed the issue when particular VM has multiple volumes
 > attached and when new OSDs are added into cinder volume ceph
 > cluster.
 > 
 > http://ceph-users.ceph.narkive.com/baW6rD2y/nfs-interaction-with-rbd
 > http://lists.ceph.com/pipermail/ceph-users-ceph.com/2013-December/026187.html

I've been talking to RedHat a bit about this and yesterday raised it on the ceph-users list. The issue is the number of socket FDs that the qemu process ends up using with volumes from various network storage interfaces, e.g., ceph rbd (but I imagine there would be similar issues with other less popular block storage cluster tech too - gluster, sheepdog, etc).


The feedback I've had for Ceph librbd is essentially that you need to allow for number of sockets equal to:
<qemu base usage> + ( <max num volumes per VM> * 2 * <num OSDs and MONs in cluster> )
And then on top of that you should set max_processes to something like double the max_files number. In practice the default soft ulimits for qemu VMs on Trusty seem to be files=1024 and processes=<something much larger>.

The utility of trying to tune these limits in the first place is highly questionable, because when the limits are hit the failure behaviour is far from graceful (in fact there are not even any useful logs at the qemu or libvirt layers). So at the end of all that I don't really think it's worth hiera-ising these settings, we should just pick appropriately large numbers and be done with it. 32k is more files that anyone will ever need... ;-)

Patch-set: 1