Update patch set 1
Patch Set 1:

> OK cool, so what is the current default for people that haven't
 > touched the file?

1024 seems to be the system default soft open files ulimit, unless that's been altered elsewhere, e.g., in /etc/security/limits.[conf,d].

I don't see any reason not to just make it 32k - for us (and I suspect we have the largest such storage cluster in the RC) that would mean we'd be able to support up to somewhere over 100 volumes attached per VM... I hope we never encounter a use-case that tries this! But even if we did, bumping up the default again would not impact anyone.

Patch-set: 1
Label: Code-Review=0