fix line breaks
The -%> tag trims the following line break which causes the next line to be
appended to current line, which is not what we want.

This way creates empty blank lines if the variable is not set. The alternative
is to have the output on a separate line from the if statement, like so:
 <% if @value -%>
 <%= @value %>
 <% end -%>

But this will do for now.

Change-Id: Id959fc80a3ad2bbf0d243b2d17017538de98f95a