Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > When Centos becomes EOL it will be removed from the list and this will be before we go to next LTS
> 
> There isn't anything stopping users from launching their own images that might be affected by this. It would be very painful for a user to debug. You could take a hard-line approach and just say we don't support you.. but ¯\_(ツ)_/¯

Another alternative is that we push this only to hypervisors that runs ephemeral flavours (e.g. m1, m2). Users get the best of both worlds, but this will require more coordination and I'm unsure the exact way to implement this.

Perhaps something can be done with the new HW refresh i.e. they only run m3 and we gradually drop HW that runs m1 and m2.

Patch-set: 3