Make API boxes aware that they are behind a proxy
This fixes a bug with pagination throwing an error when doing a server
list with >1000 servers using openstacksdk.

Our API boxes are being a HTTPS proxy, and the proxy server sets
'X-Forwarded-Proto: https' header when forwarding to the servers over
HTTP. The API servers must be made aware of this, so that any links
returned by them will be in format 'https://<server>/' instead of plain
'http'.

Change-Id: I992a907118ad9aaf3dd33bd8e32974b2b41395f1