Add limit_tenants_to_placement_aggregate to nova.conf
This setting causes the scheduler to look up a host aggregate with the
metadata key of filter_tenant_id set to the project of an incoming
request, and request results from placement be limited to that aggregate.
Multiple tenants may be added to a single aggregate by appending a serial
number to the key, such as filter_tenant_id:123.

The matching aggregate UUID must be mirrored in placement for proper
operation. If no host aggregate with the tenant id is found, or that
aggregate does not match one in placement, the result will be the same
as not finding any suitable hosts for the request.

Change-Id: I4e7b1a338947e03ce42a435d7d6837cd21be7a8e