Add support for image_cache options
This patch introduces a new class named nova::compute::image_cache,
to support parameters in image_cache sections, which were recently
added into nova[1].

[1] 828e8047e5c8651ea757bda7922670889d5e8818

(Nectar) Backported from upstream, but We patch the image_cache.pp file
to still configure the train options. This is because image_cache section
is only introduced since ussuri, but in this way we don't need to change
the defined hiera when we abandon this private path and use upstream
vitoria/wallaby code.

Change-Id: I61d81667e4c58b62c453e37072d9aef461923b46