Use deprecated devices/enabled_vgpu_types
An upstream change in both nova[1] and puppet-nova[1] at Xena changed
the config 'devices/enabled_vgpu_types' to 'devices/enabled_mdev_types'.

This means that we cannot have <Xena version of compute nodes running
when we upgrade to Xena puppet module.

Put back the old code so that we can continue supporting older compute
nodes after we have upgraded puppet modules to Xena. Also remove
deprecation warning since it is not valid for versions <Xena.

This patch can be removed when compute nodes are at Xena.

[1] https://review.opendev.org/c/openstack/nova/+/801607
[2] https://review.opendev.org/c/openstack/puppet-nova/+/810429

Change-Id: I19c7ad4b0874b3bb4fefe1fcc689e4fe9e0cbd8b